refactor: legacy modal wrapper
