+++ features/stake/stake-form/stake-form-modal.tsx
@@ -1,22 +1,37 @@
-//import { TxStageModal } from 'shared/components/tx-stage-modal';
+import { useTransactionModal } from 'shared/transaction-modal';
+import { convertTxStageToLegacy } from 'features/wsteth/shared/utils/convertTxModalStageToLegacy';
+import { TxStageModal } from 'shared/components';
+import { formatBalance } from 'utils';
+import { TX_OPERATION as TX_OPERATION_LEGACY } from 'shared/components/tx-stage-modal';
+import { useStakeFormData } from './stake-form-context';
 
 export const StakeFormModal = () => {
-  return null;
-  // return (
-  // <TxStageModal
-  //   open={txModalOpen}
-  //   onClose={closeTxModal}
-  //   txStage={txStage}
-  //   txOperation={TX_OPERATION.STAKING}
-  //   txHash={txHash}
-  //   amount={inputValue}
-  //   amountToken="ETH"
-  //   willReceiveAmount={inputValue}
-  //   willReceiveAmountToken="stETH"
-  //   balance={stethBalance.data}
-  //   balanceToken="stETH"
-  //   failedText={txModalFailedText}
-  //   onRetry={() => formRef.current?.requestSubmit()}
-  // />
-  // );
+  const { etherBalance } = useStakeFormData();
+  const {
+    dispatchModalState,
+    onRetry,
+    amount,
+    txHash,
+    txStage,
+    isModalOpen,
+    errorText,
+  } = useTransactionModal();
+
+  return (
+    <TxStageModal
+      open={isModalOpen}
+      onClose={() => dispatchModalState({ type: 'close_modal' })}
+      txStage={convertTxStageToLegacy(txStage)}
+      txOperation={TX_OPERATION_LEGACY.STAKING}
+      txHash={txHash}
+      amount={amount ? formatBalance(amount) : ''}
+      amountToken="ETH"
+      willReceiveAmount={amount ? formatBalance(amount) : undefined}
+      willReceiveAmountToken="stETH"
+      balance={etherBalance}
+      balanceToken="stETH"
+      failedText={errorText}
+      onRetry={() => onRetry?.()}
+    />
+  );
 };
