+++ features/stake/stake-form/wallet/limit-meter/icons.tsx
