+++ features/stake/stake-form/wallet/styles.tsx
