+++ shared/components/index.ts
@@ -1,5 +1,4 @@
 export { Layout } from './layout/layout';
-export { PopupWrapper } from './popup-wrapper/popup-wrapper';
 export { Faq } from './faq/faq';
 export { Header } from './header/header';
 export { Footer } from './footer/footer';
